pyproject.toml
[project]
name = "appsdk"
version = "0.1.0"
description = "Sample OpenAI App SDK project"
authors = [{ name = "Example", email = "example@example.com" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "python-dotenv>=1.0.1",
    "typer>=0.12.3",
    "uvicorn>=0.30.0",
]

[project.scripts]
appsdk = "app.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
